+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stcland/storybook
 
+## 0.0.10
+
+### Patch Changes
+
+- Added Accordion to STCLandia.
+- Updated dependencies
+  - @stcland/components@0.0.6
+
 ## 0.0.9
 
 ### Patch Changes